+++ scripts/macros/spells/spiritOfDeath.js
@@ -1,5 +1,5 @@
 import {Summons} from '../../lib/summons.js';
-import {actorUtils, compendiumUtils, constants, effectUtils, errors, genericUtils, itemUtils, workflowUtils} from '../../utils.js';
+import {activityUtils, actorUtils, compendiumUtils, constants, effectUtils, errors, genericUtils, itemUtils, workflowUtils} from '../../utils.js';
 async function use({workflow}) {
     let concentrationEffect = effectUtils.getConcentrationEffect(workflow.actor, workflow.item);
     let sourceActor = await compendiumUtils.getActorFromCompendium(constants.packs.summons, 'CPR - Reaper Spirit');
@@ -11,8 +11,8 @@ async function use({workflow}) {
     let numAttacks = Math.floor(spellLevel / 2);
     let multiAttackFeatureData = await Summons.getSummonItem('Multiattack (Reaper Spirit)', {}, workflow.item, {translate: genericUtils.format('CHRISPREMADES.CommonFeatures.Multiattack', {numAttacks}), identifier: 'spiritOfDeathMultiattack'});
     let reapingScytheFeatureData = await Summons.getSummonItem('Reaping Scythe (Reaper Spirit)', {}, workflow.item, {translate: 'CHRISPREMADES.Macros.SpiritOfDeath.ReapingScythe', identifier: 'spiritOfDeathReapingScythe', flatAttack: true, damageBonus: spellLevel});
-    let hauntCreatureFeatureData = await Summons.getSummonItem('Haunt Creature (Reaper Spirit)', {}, workflow.item, {translate: 'CHRISPREMADES.Macros.SpiritOfDeath.HauntCreature', identifier: 'spiritOfDeathHauntCreature'});
-    if (!multiAttackFeatureData || !reapingScytheFeatureData) {
+    let hauntCreatureFeatureData = await Summons.getSummonItem('Haunt Creature (Reaper Spirit)', {}, workflow.item, {translate: 'CHRISPREMADES.Macros.SpiritOfDeath.HauntCreature', identifier: 'spiritOfDeathHauntCreature', flatDC: true});
+    if (!multiAttackFeatureData || !reapingScytheFeatureData || !hauntCreatureFeatureData) {
         errors.missingPackItem();
         if (concentrationEffect) await genericUtils.remove(concentrationEffect);
         return;
@@ -69,27 +69,22 @@ async function attack({workflow}) {
     if (workflow.targets.size !== 1 || workflow.advantage) return;
     let effect = effectUtils.getEffectByIdentifier(workflow.actor, 'spiritOfDeathHauntCreature');
     if (!effect) {
-        workflow.aborted = true;   
+        workflow.aborted = true;
+        genericUtils.notify('CHRISPREMADES.Macros.SpiritOfDeath.InvalidTarget', 'info');
         return;
     }
     let {targets: validTargetUuids} = effect.flags['chris-premades'].spiritOfDeathHauntCreature;
     if (!validTargetUuids.includes(workflow.targets.first().document.uuid)) {
         workflow.aborted = true;
+        genericUtils.notify('CHRISPREMADES.Macros.SpiritOfDeath.InvalidTarget', 'info');
         return;
     }
     workflow.advantage = true;
     workflow.attackAdvAttribution.add(genericUtils.translate('DND5E.Advantage') + ': ' + effect.name);
 }
 async function late({workflow}) {
-    let effect = effectUtils.getEffectByIdentifier(workflow.actor, 'summonedEffect');
-    let originItem = fromUuidSync(effect?.origin);
-    let sourceActor = originItem?.actor;
-    if(!sourceActor) return;
-    let featureData = await compendiumUtils.getItemFromCompendium(constants.packs.summonFeatures, 'Haunt Creature: Haunt', {object: true, getDescription: true, translate: 'CHRISPREMADES.Macros.SpiritOfDeath.HauntCreatureHaunt', identifier: 'spiritOfDeathHauntCreatureHaunt', flatDC: sourceActor.system.attributes.spelldc});
-    if (!featureData) {
-        errors.missingPackItem();
-        return;
-    }
+    let effect = effectUtils.getEffectByIdentifier(workflow.actor, 'spiritOfDeathHauntCreature');
+    if (effect) await genericUtils.remove(effect);
     let targetEffectData = {
         name: genericUtils.translate('CHRISPREMADES.Macros.SpiritOfDeath.Haunted'),
         img: workflow.item.img,
@@ -120,18 +115,17 @@ async function late({workflow}) {
     await Promise.all(workflow.targets.map(async target => {
         if (target.actor) await effectUtils.createEffect(target.actor, targetEffectData, {strictlyInterdependent: true, parentEntity: casterEffect, identifier: 'spiritOfDeathHauntCreatureHaunted'});
     }));
-    await itemUtils.createItems(workflow.actor, [featureData], {parentEntity: casterEffect});
 }
 async function turnStart({trigger: {entity: effect, token, target}}) {
     let hauntedEffect = effectUtils.getEffectByIdentifier(target.actor, 'spiritOfDeathHauntCreatureHaunted');
     if (!hauntedEffect) return;
-    let feature = itemUtils.getItemByIdentifier(token.actor, 'spiritOfDeathHauntCreatureHaunt');
+    let feature = activityUtils.getActivityByIdentifier(fromUuidSync(effect.origin), 'hauntCreatureSave', {strict: true});
     if (!feature) return;
-    await workflowUtils.syntheticItemRoll(feature, [target]);
+    await workflowUtils.syntheticActivityRoll(feature, [target]);
 }
 export let spiritOfDeath = {
     name: 'Spirit of Death',
-    version: '1.0.50',
+    version: '1.1.1',
     midi: {
         item: [
             {
@@ -169,6 +163,7 @@ export let spiritOfDeath = {
         {
             value: 'animation',
             label: 'CHRISPREMADES.Config.SpecificAnimation',
+            i18nOption: 'CHRISPREMADES.Summons.CreatureNames.ReaperSpirit',
             type: 'select',
             default: 'shadow',
             category: 'animation',
@@ -197,7 +192,8 @@ export let spiritOfDeathHauntCreature = {
             {
                 pass: 'rollFinished',
                 macro: late,
-                priority: 50
+                priority: 50,
+                activities: ['hauntCreatureHaunt']
             }
         ]
     }